+++ TP3/grille.h
@@ -10,8 +10,8 @@ public:
     int GetColonnes() const;
     void InitialiserGrille();
     void AfficherGrille() const;
-    virtual bool CaseVide(const size_t ligne, const size_t colonne) const;
-    void DeposerMorpion(const size_t ligne, const size_t colonne, const char jeton);
+    virtual bool CaseVide(const int ligne, const int colonne) const;
+    void DeposerMorpion(const int ligne, const int colonne, const char jeton);
     virtual bool EstLigne(const char jeton) const;
     virtual bool EstColonne(const char jeton) const;
     virtual bool EstDiagonale(const char jeton) const;
@@ -29,8 +29,8 @@ class GrillePuissance4 : public Grille
 public:
     GrillePuissance4(){};
     GrillePuissance4(const int _nbLignes, const int _nbColonnes) : Grille(_nbLignes,_nbColonnes){};
-    bool DeposerPuissance4(const size_t colonne, const char jeton);
-    int CompteJeton(const size_t ligne, const size_t colonne, const int dirLigne, const int dirColonne) const;
+    bool DeposerPuissance4(const int colonne, const char jeton);
+    int CompteJeton(const int ligne, const int colonne, const int dirLigne, const int dirColonne) const;
     bool EstLigne(const char jeton) const override;
     bool EstColonne(const char jeton) const override;
     bool EstDiagonale(const char jeton) const override;
